requirements.txt
Django==3.2
psycopg2-binary==2.9.1
djangorestframework==3.12.4
